pyproject.toml
[project]
name = "qwirkle"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Kevin McWhirter", email = "klmcw@yahoo.com"}
]
license = { text = "MIT" }
readme = "README.md"

requires-python = "==3.12.*"
dependencies = [
    "pygame-ce>=2.4.0"
]


[tool.pdm]
distribution = false


[tool.pdm.scripts]
create_venv = { shell = "pdm venv remove -y in-project; pdm install --prod" }
install_dev = { shell = "pdm install --dev" }
start = "python -m qwirkle"
clean = { shell = "find . -type d \\( -name __pycache__ -o -name .pytest_cache \\) -exec rm -fr {} \\;" }
test = "pytest -vv"


[tool.pdm.dev-dependencies]
dev = [
    "autopep8>=2.0.4",
    "pytest>=8.0.0",
]
